[FEAT] SAA 아키텍처링 및 MVVM 패턴으로 마이그레이션
