+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2026-04-24T04:21:01.216056Z">
+        <DropdownSelection timestamp="2026-04-30T05:31:33.589025800Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="PhysicalDevice" identifier="serial=R3CW20J7V4X" />
+              <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\TehoonAhn\.android\avd\Pixel_9.avd" />
             </handle>
           </Target>
         </DropdownSelection>
