+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.LETSSOPT">
         <activity
-            android:name=".ui.login.LoginActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.LETSSOPT">
@@ -22,18 +22,6 @@
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".ui.signup.SignUpActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.LETSSOPT">
-        </activity>
-        <activity
-            android:name=".ui.home.MainActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.LETSSOPT">
         </activity>
     </application>
