+++ app/src/main/java/com/example/letssopt/ui/signup/SignUpScreen.kt
@@ -1,0 +1,259 @@
+package com.example.letssopt.ui.signup
+
+import android.content.Context
+import android.widget.Toast
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.consumeWindowInsets
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Snackbar
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusDirection
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.example.letssopt.ui.components.DefaultButton
+import com.example.letssopt.ui.components.DefaultTextField
+import com.example.letssopt.ui.theme.AsBg
+import com.example.letssopt.ui.theme.AsPrimary
+import com.example.letssopt.ui.theme.AsSecondaryText
+import com.example.letssopt.ui.theme.AsSurface
+import com.example.letssopt.ui.theme.AsWhite
+import com.example.letssopt.ui.theme.LETSSOPTTheme
+import kotlinx.coroutines.launch
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object SIGNUP
+
+@Composable
+fun SignupScreen(
+    modifier: Modifier = Modifier,
+    viewModel: SignUpViewModel = viewModel(),
+    onSignUpComplete: () -> Unit = {}
+) {
+    val context = LocalContext.current
+
+    val focusManager = LocalFocusManager.current
+    val scrollState = rememberScrollState()
+
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
+
+    val scope = rememberCoroutineScope()
+    val snackbarHostState = remember { SnackbarHostState() }
+
+    val onShowSnack: (String) -> Unit = { message ->
+        scope.launch {
+            snackbarHostState.showSnackbar(message)
+        }
+    }
+
+    Scaffold(
+        snackbarHost = {
+            SnackbarHost(
+                modifier = Modifier
+                    .navigationBarsPadding()
+                    .imePadding(),
+                hostState = snackbarHostState
+            ) {
+                Snackbar(
+                    snackbarData = it,
+                    containerColor = AsSurface,
+                    contentColor = AsWhite
+                )
+            }
+        },
+        modifier = modifier.fillMaxSize()
+    ) { innerPadding ->
+        //전반적인 레이아웃
+        Column(
+            modifier = modifier
+                .fillMaxSize()
+                .background(color = AsBg)
+                .padding(innerPadding)
+                .consumeWindowInsets(innerPadding)
+                .padding(20.dp)
+                .imePadding(),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            //위쪽 컴포넌트 정렬을 위한 2차 레이아웃 wrapper
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .verticalScroll(scrollState)
+                    .weight(1f),
+                horizontalAlignment = Alignment.Start,
+                verticalArrangement = Arrangement.Top
+            ) {
+                //로고
+                Text(
+                    modifier = Modifier
+                        .padding(top = 40.dp)
+                        .align(Alignment.CenterHorizontally),
+                    style = MaterialTheme.typography.titleLarge.copy(
+                        fontSize = 36.sp,
+                        color = AsPrimary
+                    ),
+                    text = "watcha"
+                )
+
+                Text(
+                    modifier = Modifier
+                        .padding(top = 20.dp),
+                    style = MaterialTheme.typography.titleLarge.copy(fontSize = 20.sp),
+                    text = "회원가입",
+                    color = AsWhite
+                )
+                //텍스트필드 섹션
+                Text(
+                    modifier = Modifier
+                        .padding(top = 28.dp),
+                    style = MaterialTheme.typography.labelSmall,
+                    text = "이메일",
+                    color = AsSecondaryText
+                )
+
+                DefaultTextField(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = 4.dp),
+                    text = uiState.textId,
+                    onValueChange = { viewModel.onChangedId(it) },
+                    hint = "이메일 주소를 입력하세요",
+                    tfVisible = true,
+                    keyboardOptions = KeyboardOptions(
+                        keyboardType = KeyboardType.Email,
+                        imeAction = ImeAction.Next
+                    ),
+                    keyboardActions = KeyboardActions(
+                        onNext = { focusManager.moveFocus(FocusDirection.Down) }
+                    )
+                )
+                Text(
+                    style = MaterialTheme.typography.labelSmall,
+                    text = "비밀번호",
+                    color = AsSecondaryText,
+                    modifier = Modifier.padding(top = 12.dp)
+                )
+
+                DefaultTextField(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = 4.dp),
+                    text = uiState.textPw,
+                    onValueChange = { viewModel.onChangedPw(it) },
+                    hint = "비밀번호를 입력하세요",
+                    tfVisible = false,
+                    keyboardOptions = KeyboardOptions(
+                        keyboardType = KeyboardType.Password,
+                        imeAction = ImeAction.Next
+                    ),
+                    keyboardActions = KeyboardActions(
+                        onNext = { focusManager.moveFocus(FocusDirection.Down) }
+                    )
+                )
+
+                Text(
+                    modifier = Modifier
+                        .padding(top = 12.dp),
+                    style = MaterialTheme.typography.labelSmall,
+                    text = "비밀번호 확인",
+                    color = AsSecondaryText
+                )
+
+                DefaultTextField(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = 4.dp),
+                    text = uiState.textCkPw,
+                    onValueChange = { viewModel.onChangedCkPw(it) },
+                    hint = "비밀번호를 다시 입력하세요",
+                    tfVisible = false,
+                    keyboardOptions = KeyboardOptions(
+                        keyboardType = KeyboardType.Password,
+                        imeAction = ImeAction.Done
+                    ),
+                    keyboardActions = KeyboardActions(
+                        onDone = { focusManager.clearFocus() }
+                    )
+                )
+            }
+            // 이쪽 부분에선 래핑해줬던 레이아웃 제거
+            DefaultButton(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = 10.dp, top = 20.dp),
+                "회원가입",
+                onClick = {
+
+                    validateSignUp(
+                        context = context,
+                        textId = uiState.textId,
+                        textPw = uiState.textPw,
+                        textPwCheck = uiState.textCkPw,
+                        viewModel = viewModel,
+                        onSignUpComplete = onSignUpComplete,
+                        onShowSnack = onShowSnack
+                    )
+
+                },
+                btEnabled = uiState.textId.isNotEmpty() &&
+                        uiState.textPw.isNotEmpty() &&
+                        uiState.textCkPw.isNotEmpty()
+            )
+        }
+    }
+}
+
+private fun validateSignUp(
+    context: Context,
+    textId: String,
+    textPw: String,
+    textPwCheck: String,
+    onSignUpComplete: () -> Unit = {},
+    onShowSnack: (String) -> Unit,
+    viewModel: SignUpViewModel
+) {
+    if (viewModel.validationCheck(textId, textPw, textPwCheck)
+    ) {
+        Toast.makeText(context, "회원가입 성공!", Toast.LENGTH_SHORT).show()
+        viewModel.onSaveAccount(textId, textPw)
+        onSignUpComplete()
+    } else {
+        onShowSnack("회원가입에 실패했습니다. 올바른 정보를 입력해주세요.")
+    }
+}
+
+
+@Preview(showBackground = true)
+@Composable
+private fun SignupContentPreview() {
+    LETSSOPTTheme {
+        SignupScreen()
+    }
+}
